Moves hbond_autocorrel and removes old hbond code (#3258)
Fixes #2739 #3259 

## Work done in this PR
* Moves hbond_autocorrel from hbonds to hydrogenbonds, stub
  will be removed in 3.0.0.
* Removes deprecated hbonds.hbond_analysis code.
* tests (note: tests were copied and we run full tests on the deprecated
  and the moved code, mainly to make sure that they can be called as
  expected; deprecation warnings are also tested)
* make hydrogenbonds.hbond_autocorrel primary entry for docs
* add entry for deprecated module at bottom of Hydrogen bonding docs
* fixed doc in hydrogenbonds.wbridge_analysis
* include the Gregoret 1991 reference (that was removed with the
  deprecated hbonds.hbond_analysis module)
