+++ package/CHANGELOG
@@ -36,7 +36,7 @@ Fixes
     removed (Issue #3209)
   * Fixed 'sphzone', 'sphlayer', 'cyzone', and 'cylayer' to return empty if the
    zone/layer is empty, consistent with 'around' (Issue #2915)
-  * A Universe created from an ROMol with no atoms returns now a Universe 
+  * A Universe created from an ROMol with no atoms returns now a Universe
     with 0 atoms (Issue #3142)
   * PDBParser will check for the presence of the chainID attribute of an
     atom group and use these values instead of just using the end of segid.
@@ -114,10 +114,10 @@ Enhancements
   * Switch GNMAnalysis to AnalysisBase (Issue #3243)
   * Adds python 3.9 support (Issue #2974, PR #3027, #3245)
   * Added an MDAnalysis shields.io badge to the README (Issue #3227, PR #3229)
-  * Added sort method to the atomgroup (Issue #2976, PR #3188) 
+  * Added sort method to the atomgroup (Issue #2976, PR #3188)
   * ITPParser now reads [ atomtypes ] sections in ITP files, used for charges
     and masses not defined in the [ atoms ] sections
-  * Add `set_dimensions` transformation class for setting constant 
+  * Add `set_dimensions` transformation class for setting constant
     box dimensions for all timesteps in trajectory (Issue #2691)
   * Added a ValueError raised when not given a gridcenter while
     providing grid dimensions to DensityAnalysis, also added
@@ -174,6 +174,8 @@ Enhancements
     checking if it can be used in parallel analysis. (Issue #2996, PR #2950)
 
 Changes
+  * Deprecated analysis.hbonds.hbond_analysis has been removed in favour of
+    analysis.hydrogenbonds.hbond_analysis (Issues #2739, #2746)
   * `GNMAnalysis`, `LinearDensity`, `PersistenceLength` and
     `AnalysisFromFunction` use the `results` attribute.
   * Fixed inaccurate docstring inside the RMSD class (Issue #2796, PR #3134) 
@@ -220,6 +222,9 @@ Changes
   * Added OpenMM coordinate and topology converters (Issue #2863, PR #2917)
 
 Deprecations
+  * The analysis.hbonds.hbond_autocorrel code has been moved to
+    analysis.hydrogenbonds.hbond_autocorrel and will be removed
+    from analysis.hbonds in 3.0.0 (PR #3258)
   * In 2.1.0 the TRZReader will default to a dt of 1.0 ps when failing to
     read it from the input TRZ trajectory.
