+++ package/MDAnalysis/analysis/hydrogenbonds/__init__.py
@@ -23,8 +23,10 @@
 
 __all__ = [
     'HydrogenBondAnalysis',
-    'WaterBridgeAnalysis'
+    'WaterBridgeAnalysis',
+    'HydrogenBondAutoCorrel', 'find_hydrogen_donors'
 ]
 
 from .hbond_analysis import HydrogenBondAnalysis
 from .wbridge_analysis import WaterBridgeAnalysis
+from .hbond_autocorrel import HydrogenBondAutoCorrel, find_hydrogen_donors
