--- package/doc/sphinx/source/documentation_pages/analysis/hbond_analysis.rst
@@ -1,1 +1,0 @@
-.. automodule:: MDAnalysis.analysis.hbonds.hbond_analysis
