+++ package/doc/sphinx/source/documentation_pages/analysis/hbond_autocorrel_deprecated.rst
@@ -1,0 +1,1 @@
+.. automodule:: MDAnalysis.analysis.hbonds.hbond_autocorrel
